debugging db, working on deletion (unlike)

+++ app/src/main/java/com/example/jinjinz/concertprev/models/Concert.java
@@ -33,7 +33,7 @@ public class Concert {
     private ArrayList<String> artists;
 
 
-    public void setDbId(int dbId) {
+    public void setDbId(long dbId) {
         this.dbId = dbId;
     }
     public void setEventName(String eventName) {
